[level 0] Title: 구슬을 나누는 경우의 수, Time: 0.06 ms, Memory: 78.2 MB -BaekjoonHub

+++ Java/프로그래머스/0/120840. 구슬을 나누는 경우의 수/구슬을 나누는 경우의 수.java
@@ -1,5 +1,5 @@
 class Solution {
-    public double solution(int balls, int share) {
+    public int solution(int balls, int share) {
         double answer = 0;
         
         double n = 1;
@@ -18,6 +18,6 @@ class Solution {
         }
         
         answer = n/(nm*m);
-        return (int)Math.round(answer);
+        return (int) Math.round(answer);
     }
 }
